requirements.txt
# 核心依赖
aiohttp>=3.9.0
websockets>=12.0
redis[hiredis]>=5.0.0

# 数学计算
numpy>=1.26.0
scipy>=1.11.0

# 配置
pyyaml>=6.0

# 可选：链上交互（Phase 2+）
# web3>=6.0.0

# 可选：OKX SDK
# okx-api>=1.0.0

# 开发
pytest>=7.4.0
pytest-asyncio>=0.23.0
